+++ packages/grid/src/GridMetricCalculator.test.ts
@@ -1,4 +1,4 @@
-import { getOrThrow, trimMap } from './GridMetricCalculator';
+import { trimMap } from './GridMetricCalculator';
 
 describe('trimMap', () => {
   function makeMap(low = 0, high = 10): Map<number, number> {
@@ -39,36 +39,5 @@ describe('trimMap', () => {
     expectResult(makeMap(0, 10), makeMap(6, 10));
     expectResult(makeMap(0, 100), makeMap(96, 100));
     expectResult(makeMap(0, 100), makeMap(51, 100), 100, 50);
-  });
-});
-
-describe('getOrThrow', () => {
-  const MAP = new Map([
-    [5, 10],
-    [6, 16],
-    [10, 50],
-    [100, 250],
-  ]);
-
-  it('gets the value if it exists', () => {
-    expect(getOrThrow(MAP, 5)).toBe(10);
-    expect(getOrThrow(MAP, 6)).toBe(16);
-    expect(getOrThrow(MAP, 10)).toBe(50);
-    expect(getOrThrow(MAP, 100)).toBe(250);
-  });
-
-  it('gets the value if it exists even if default provided', () => {
-    expect(getOrThrow(MAP, 5, 7)).toBe(10);
-    expect(getOrThrow(MAP, 6, 7)).toBe(16);
-    expect(getOrThrow(MAP, 10, 7)).toBe(50);
-    expect(getOrThrow(MAP, 100, 7)).toBe(250);
-  });
-
-  it('throws if no value set', () => {
-    expect(() => getOrThrow(MAP, 0)).toThrow();
-  });
-
-  it('returns default value if provided', () => {
-    expect(getOrThrow(MAP, 0, 7)).toBe(7);
   });
 });