+++ packages/grid/src/mouse-handlers/GridRowMoveMouseHandler.ts
@@ -1,6 +1,6 @@
+import { getOrThrow } from '@deephaven/utils';
 import { EventHandlerResult } from '../EventHandlerResult';
 import Grid from '../Grid';
-import { getOrThrow } from '../GridMetricCalculator';
 import GridMouseHandler from '../GridMouseHandler';
 import GridUtils, { GridPoint } from '../GridUtils';