+++ packages/iris-grid/src/IrisGridTheme.ts
@@ -71,6 +71,7 @@ const theme: Partial<IrisGridThemeType> = Object.freeze({
   scrollBarActiveSelectionTickColor:
     IrisGridTheme['scroll-bar-active-selection-tick-color'],
   textColor: IrisGridTheme['text-color'],
+  hyperlinkColor: IrisGridTheme['hyperlink-color'],
   positiveNumberColor: IrisGridTheme['positive-number-color'],
   negativeNumberColor: IrisGridTheme['negative-number-color'],
   zeroNumberColor: IrisGridTheme['zero-number-color'],